backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
geoalchemy2==0.15.2
pydantic-settings==2.7.1
alembic==1.14.1
pandas==2.2.3
psycopg2-binary==2.9.10
python-dotenv==1.0.1
